pyproject.toml
[tool.poetry]
name = "drf_shoestore_api"
version = "0.1.0"
description = ""
authors = ["Justin Miller <indymiller@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.7"
djangorestframework = "^3.11.0"
django-cors-headers = "^3.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
